E11: frontier max-keep, shared send window, single input config
- update_buffer only advances last_remote_tick (max-keep): stale
  out-of-order messages no longer move the confirmed frontier backward
  (protects check_rollback/LastConfirmedInput); adds stale-frontier test.
- Shared send_window() borrow-window builder in the core crate; all three
  encoders (native/leafwing/BEI) walk it instead of per-tick gets.
  Zero-clone (borrows only), wire-identical, backends keep their own
  start selection and encoding.
- ServerInputConfig removed; ServerInputPlugin stamps rebroadcast_inputs
  onto the single InputConfig resource (inserting a default first).
- MAX_INPUT_PAST_TICKS 128 -> 64: same window as the ring and lookahead.
- Native: drop dead IntoIterator impl (duplicated
  get_snapshots_from_message; the trait deliberately avoids IntoIterator).
- Leafwing: refresh stale SameAsPrecedent-buffer test comments.
- num_ticks u32 -> usize across the trait, backends, and client.

+++ crates/inputs/inputs/src/client.rs
@@ -791,11 +791,11 @@ fn prepare_input_message<S: ActionStateSequence>(
 
     // Send redundant inputs so that if a packet is lost, we can still recover.
     // The size of the input bundle scales with `send_interval`.
-    let mut num_ticks: u32 = ((input_config.send_interval.as_nanos() / tick_duration.as_nanos())
+    let mut num_ticks: usize = ((input_config.send_interval.as_nanos() / tick_duration.as_nanos())
         + 1)
     .try_into()
     .unwrap();
-    num_ticks *= input_config.packet_redundancy as u32;
+    num_ticks *= input_config.packet_redundancy as usize;
     let mut message = InputMessage::<S>::new(tick);
     for (entity, input_buffer, pre_spawned, controlled_by) in input_buffer_query.iter() {
         if !route.accepts_local_target(controlled_by) {
